+++ src/components/restaurantManagement/restaurants/RestaurantAddEmp.tsx
@@ -5,7 +5,6 @@ import { useTranslation } from "react-i18next";
 import { useValidationSchemas } from "../../../hooks/useValidationSchema";
 import { fetchGET, fetchPOST } from "../../../services/APIconn";
 import ErrorMes from "../../reusableComponents/ErrorMessage";
-import { RestaurantDataProps } from "../../../services/interfaces";
 import { EmploymentType } from "../../../services/types";
 
 const initialValues = {
